ast: Moved partial application code from `Call` to `ParsedCall`.
